refactor/log: 충돌 수정
